Initial attempt at porting Scientist
https://github.com/github/scientist

There is a Java implementation but I thought we could do it more
simply and clearly in Scala.

I've also borrowed the 'ignore' behaviour, which looks very useful (for
example, ignoring results less than 1 minute old to account for
replication lag) but doesn't exist in the Java port.

And added, too, a default publish option using a library to diff case
classes.